[Chore] 뷰컨트롤러 이름 변경 (#46)
